+++ tests/testthat/test-tp.R
@@ -218,6 +218,7 @@ test_that("basic usage", {
 
 TPFormLib <- genFormulaTPLibrary(patRoonData::suspectsPos)
 TPFormLibG2 <- genFormulaTPLibrary(patRoonData::suspectsPos, generations = 2)
+genTPsFormLibG2 <- generateTPs("library_formula", parents = patRoonData::suspectsPos[1:4, ], TPLibrary = TPFormLibG2)
 test_that("genFormulaTPLibrary works", {
     checkmate::expect_data_table(TPFormLib, any.missing = FALSE)
     checkmate::expect_names(names(TPFormLib), must.include = c("parent_name", "parent_formula", "parent_neutralMass",
@@ -227,6 +228,7 @@ test_that("genFormulaTPLibrary works", {
     checkmate::expect_subset(TPFormLibG2[generation == 2]$parent_name, TPFormLibG2[generation == 1]$TP_name)
     
     expect_length(generateTPs("library_formula", TPLibrary = TPFormLib), nrow(TPFormLib))
+    expect_setequal(as.data.table(genTPsFormLibG2)$generation, c(1, 2))
 })
 
 TPsCons <- consensus(TPsLibScr, TPsBTScr)