+++ NEWS.md
@@ -3,6 +3,8 @@
 * Fixed: The `removeTPIsomers` filter for `transformationProductsStructure` didn't actually apply the `removeDuplicates` filter.
 * Fixed: if `updateScores=TRUE` for the methods `addFormulaScoring()`, `predictRespFactors()` and `predictTox()` then `NaN` scores could be introduced if the `formulaScore` is zero.
 * Fixed: if `updateScores=TRUE` for the method `addFormulaScoring()` then the `score` would be updated twice.
+* Fixed: `generateTPsLibrary()`/`generateTPsLibraryFormula()` specifying >1 generation did not yield in additional TP searches if `parents!=NULL`
+* `generateTPsLibrary()`/`generateTPsLibraryFormula()` any TPs that are equal to the parent (i.e. common with `generations>1`) are now removed
 
 
 # patRoon 2.3.2